langchain_esra/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-experimental>=0.2.0
langchain-google-genai>=0.2.0
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.0.0
tiktoken>=0.4.0
faiss-cpu>=1.7.0
protobuf>=3.20.0
mysql-connector-python>=9.0.0
pymysql>=1.0.0
sentence-transformers>=2.0.0
chromadb>=0.3.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pandas>=1.0.0